requirements.txt
# cs_prac — seed.yaml 앱 (개인화 애니 추천)
streamlit>=1.30
python-dotenv>=1.0
langchain-openai>=1.0
langchain-core>=1.0
langgraph>=0.2
httpx>=0.27
duckduckgo-search>=6.0
pydantic>=2.0
